Move main loop into run.ts
