+++ modules/identity-backfill-batch/src/functions/index.ts
@@ -13,7 +13,6 @@ export * from './filterRows';
 export * from './formatDuration';
 export * from './loadConfig';
 export * from './loadState';
-export * from './main';
 export * from './parseArgs';
 export * from './parseLine';
 export * from './pickEmail';
